sqdq
sqdq

+++ Assets/Scripts/Character.cs
@@ -5,12 +5,12 @@ using Cinemachine;
 using UnityEngine;
 
 public class Character : MonoBehaviour {
-    private int _size = 2;
+    private float _size = 2;
     private Vector3 GrowthMulti = new Vector3(1, 1, 1);
     private float CoolDown = 3.0f;
     private float CoolTimer = 0;
 
-    public int Size {
+    public float Size {
         get { return _size; }
         set {
             _size = value;
@@ -21,7 +21,7 @@ public class Character : MonoBehaviour {
         }
     }
 
-    private void Grow(int size) {
+    private void Grow(float size) {
         if (CoolTimer > CoolDown) {
             CoolTimer = 0;
             transform.localScale = GrowthMulti * size;
